Extract ObjectDetailBody to a separate file
