+++ frontend/src/metabase/visualizations/components/ObjectDetail/ObjectDetailBody.styled.tsx
@@ -1,0 +1,7 @@
+import styled from "@emotion/styled";
+
+export const ObjectDetailBodyWrapper = styled.div`
+  font-size: 1rem;
+  flex: 1;
+  overflow-y: auto;
+`;
