+++ frontend/src/metabase/visualizations/components/ObjectDetail/ObjectDetailBody.tsx
@@ -1,0 +1,61 @@
+import type { DatasetData, VisualizationSettings } from "metabase-types/api";
+import type ForeignKey from "metabase-lib/metadata/ForeignKey";
+
+import { ObjectDetailBodyWrapper } from "./ObjectDetailBody.styled";
+import { DetailsTable } from "./ObjectDetailsTable";
+import { Relationships } from "./ObjectRelationships";
+import type { OnVisualizationClickType } from "./types";
+
+export interface ObjectDetailBodyProps {
+  data: DatasetData;
+  objectName: string;
+  zoomedRow: unknown[];
+  settings: VisualizationSettings;
+  hasRelationships: boolean;
+  onVisualizationClick: OnVisualizationClickType;
+  visualizationIsClickable: (clicked: unknown) => boolean;
+  tableForeignKeys?: ForeignKey[];
+  tableForeignKeyReferences?: {
+    [key: number]: { status: number; value: number };
+  };
+  followForeignKey?: (fk: ForeignKey) => void;
+}
+
+export function ObjectDetailBody({
+  data,
+  objectName,
+  zoomedRow,
+  settings,
+  hasRelationships = false,
+  onVisualizationClick,
+  visualizationIsClickable,
+  tableForeignKeys,
+  tableForeignKeyReferences,
+  followForeignKey,
+}: ObjectDetailBodyProps): JSX.Element {
+  const showRelationships =
+    hasRelationships &&
+    tableForeignKeys &&
+    tableForeignKeyReferences &&
+    followForeignKey;
+
+  return (
+    <ObjectDetailBodyWrapper>
+      <DetailsTable
+        data={data}
+        zoomedRow={zoomedRow}
+        settings={settings}
+        onVisualizationClick={onVisualizationClick}
+        visualizationIsClickable={visualizationIsClickable}
+      />
+      {showRelationships && (
+        <Relationships
+          objectName={objectName}
+          tableForeignKeys={tableForeignKeys}
+          tableForeignKeyReferences={tableForeignKeyReferences}
+          foreignKeyClicked={followForeignKey}
+        />
+      )}
+    </ObjectDetailBodyWrapper>
+  );
+}
